add filter, order by clauses for aggregate expressions
